# api/openapi.yaml
openapi: 3.0.3
info:
  title: Support Desk Backend API
  version: 1.0.0
  description: API for support tickets, attachments, and ticket status updates.
servers:
  - url: http://localhost:3000
paths:
  /api/health:
    get:
      summary: Health check
      responses:
        '200':
          description: API health status
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      status:
                        type: string
                        example: ok

  /api/tickets:
    get:
      summary: List tickets
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [open, in_progress, waiting_on_customer, resolved, closed]
        - in: query
          name: priority
          schema:
            type: string
            enum: [low, medium, high, urgent]
        - in: query
          name: sortBy
          schema:
            type: string
            enum: [updatedAt, createdAt, priority]
            default: updatedAt
        - in: query
          name: sortOrder
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Paginated ticket list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TicketWithAttachments'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'

    post:
      summary: Create ticket with optional image attachments
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - subject
                - customerName
                - customerEmail
                - description
              properties:
                subject:
                  type: string
                  example: Unable to download invoice
                customerName:
                  type: string
                  example: Alex Nguyen
                customerEmail:
                  type: string
                  format: email
                  example: alex.nguyen@example.com
                priority:
                  type: string
                  enum: [low, medium, high, urgent]
                  default: medium
                categoryId:
                  type: integer
                  nullable: true
                  example: 1
                description:
                  type: string
                  example: Download button does nothing on billing page.
                attachments:
                  type: array
                  description: Optional image files (png/jpeg/webp), max 5MB each.
                  items:
                    type: string
                    format: binary
      responses:
        '201':
          description: Ticket created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/TicketWithAttachments'
        '400':
          $ref: '#/components/responses/ValidationError'

  /api/tickets/{id}:
    get:
      summary: Get ticket details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Ticket details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/TicketWithAttachments'
        '404':
          description: Ticket not found

    patch:
      summary: Update ticket status/priority
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [open, in_progress, waiting_on_customer, resolved, closed]
                priority:
                  type: string
                  enum: [low, medium, high, urgent]
      responses:
        '200':
          description: Ticket updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                      status:
                        type: string
                      priority:
                        type: string
                      updatedAt:
                        type: string
                        format: date-time
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          description: Ticket not found

components:
  schemas:
    Attachment:
      type: object
      properties:
        id:
          type: string
        ticketId:
          type: string
        originalName:
          type: string
        fileName:
          type: string
        mimeType:
          type: string
          enum: [image/png, image/jpeg, image/webp]
        size:
          type: integer
        path:
          type: string
          example: /uploads/tickets/ticket-abc123.png
        createdAt:
          type: string
          format: date-time

    TicketWithAttachments:
      type: object
      properties:
        id:
          type: string
        publicId:
          type: string
        subject:
          type: string
        customerName:
          type: string
        customerEmail:
          type: string
          format: email
        priority:
          type: string
          enum: [low, medium, high, urgent]
        status:
          type: string
          enum: [open, in_progress, waiting_on_customer, resolved, closed]
        description:
          type: string
        categoryId:
          type: integer
          nullable: true
        categoryName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'

    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer

    ErrorPayload:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties:
                type: string

  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorPayload'